Refactor wait-for-migration command
