+++ src/Octoshift/Contracts/ISourceGithubApiFactory.cs
@@ -1,0 +1,7 @@
+namespace OctoshiftCLI.Contracts;
+
+public interface ISourceGithubApiFactory
+{
+    GithubApi Create(string apiUrl = null, string sourcePersonalAccessToken = null);
+    GithubApi CreateClientNoSsl(string apiUrl = null, string sourcePersonalAccessToken = null);
+}
