+++ src/Octoshift/Contracts/ITargetGithubApiFactory.cs
@@ -1,0 +1,6 @@
+namespace OctoshiftCLI.Contracts;
+
+public interface ITargetGithubApiFactory
+{
+    GithubApi Create(string apiUrl = null, string targetPersonalAccessToken = null);
+}
