+++ src/Octoshift/Extensions/StringExtensions.cs
@@ -17,5 +17,7 @@ namespace OctoshiftCLI.Extensions
         public static ulong? ToULongOrNull(this string s) => ulong.TryParse(s, out var result) ? result : null;
 
         public static string ReplaceInvalidCharactersWithDash(this string s) => s.HasValue() ? Regex.Replace(s, @"[^\w.-]+", "-", RegexOptions.Compiled | RegexOptions.CultureInvariant) : string.Empty;
+
+        public static string SanitizeForLog(this string s) => s?.ToUpper().Replace("-", " ");
     }
 }
