+++ src/OctoshiftCLI.Tests/gei/Commands/DownloadLogsCommandTests.cs
@@ -1,7 +1,7 @@
 using System.Threading.Tasks;
 using FluentAssertions;
 using Moq;
-using OctoshiftCLI.GithubEnterpriseImporter;
+using OctoshiftCLI.Contracts;
 using OctoshiftCLI.GithubEnterpriseImporter.Commands;
 using Xunit;
