+++ src/ado2gh/Commands/GenerateMannequinCsvCommand.cs
@@ -82,7 +82,7 @@ namespace OctoshiftCLI.AdoToGithub.Commands
                 _log.LogInformation("INCLUDING RECLAIMED");
             }
 
-            var githubApi = _githubApiFactory.Create(personalAccessToken: githubPat);
+            var githubApi = _githubApiFactory.Create(targetPersonalAccessToken: githubPat);
 
             var githubOrgId = await githubApi.GetOrganizationId(githubOrg);
             var mannequins = await githubApi.GetMannequins(githubOrgId);
