+++ src/gei/Commands/GenerateMannequinCsvCommand.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Octoshift;
+using OctoshiftCLI.Contracts;
 
 namespace OctoshiftCLI.GithubEnterpriseImporter.Commands
 {
