+++ src/gei/Commands/GrantMigratorRoleCommand.cs
@@ -2,6 +2,7 @@ using System;
 using System.CommandLine;
 using System.CommandLine.Invocation;
 using System.Threading.Tasks;
+using OctoshiftCLI.Contracts;
 
 namespace OctoshiftCLI.GithubEnterpriseImporter.Commands
 {
