--- src/gei/ISourceGithubApiFactory.cs
@@ -1,8 +1,0 @@
-namespace OctoshiftCLI.GithubEnterpriseImporter
-{
-    public interface ISourceGithubApiFactory
-    {
-        GithubApi Create(string apiUrl = null, string sourcePersonalAccessToken = null);
-        GithubApi CreateClientNoSsl(string apiUrl = null, string sourcePersonalAccessToken = null);
-    }
-}
