--- src/gei/ITargetGithubApiFactory.cs
@@ -1,7 +1,0 @@
-namespace OctoshiftCLI.GithubEnterpriseImporter
-{
-    public interface ITargetGithubApiFactory
-    {
-        GithubApi Create(string apiUrl = null, string targetPersonalAccessToken = null);
-    }
-}
